pyproject.toml
[project]
name = "longmemeval-investigation"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "python-dotenv",
    "litellm",
    "pandas",
    "datasets",
    "huggingface-hub",
]

[project.optional-dependencies]
dev = [
    "ipykernel",
]
